Updated demo, improved facebook docs
